refactor(sandbox): update IconButton styles and cleanup Vite plugin
