+++ packages/headless-styles/sandbox/plugins/css-to-ts/convert-inputs/convert-inputs.js
@@ -1,5 +1,5 @@
-import path from 'path'
-import fs from 'fs'
+import { readFileSync, readdirSync, statSync } from 'node:fs'
+import { basename, join } from 'node:path'
 import convertToJS from '../convert-rules/convertToJS'
 import reverseMediaQueries from '../utils/mediaReverse'
 
@@ -12,27 +12,25 @@ export const convertStringToJson = (input, mediaReverse) => {
 }
 
 export const convertFileToJson = (inputFile, mediaReverse) => {
-  const css = fs.readFileSync(inputFile, 'utf8')
+  const css = readFileSync(inputFile, 'utf8')
 
   let contents = convertToJS(css)
   if (mediaReverse) {
     contents = reverseMediaQueries(contents)
   }
-  const filename = path.basename(inputFile, '.css')
+  const filename = basename(inputFile, '.css')
   return { contents, filename }
 }
 
 export const convertDirToJson = (inputLocation, mediaReverse) => {
-  const files = fs.readdirSync(inputLocation)
+  const files = readdirSync(inputLocation)
   return files
     .filter((file) => {
-      if (fs.statSync(path.join(inputLocation, file)).isDirectory()) {
+      if (statSync(join(inputLocation, file)).isDirectory()) {
         console.warn(`Nested directories not supported, skipping ${file}`)
         return false
       }
       return true
     })
-    .map((file) =>
-      convertFileToJson(path.join(inputLocation, file), mediaReverse)
-    )
+    .map((file) => convertFileToJson(join(inputLocation, file), mediaReverse))
 }
