+++ packages/headless-styles/sandbox/plugins/css-to-ts/convert-rules/convertToJS.js
@@ -1,24 +1,20 @@
 const css = require('css')
 import addProperty from '../utils/addProperty'
+import composes from './composes'
 import fontface from './fontface'
 import keyframes from './keyframes'
 import standard from './standard'
 
-function checkForNestedSelectors(propToCheck) {
-  if (typeof propToCheck === 'string') {
-    return propToCheck
+export default function convertToJS(input) {
+  try {
+    // Parse CSS string into rules array
+    const parsedCss = css.parse(input)
+    const { rules } = parsedCss.stylesheet
+    const initial = convertRules(rules)
+    return composes(initial)
+  } catch (err) {
+    throw new Error(`Invalid CSS input: ${err}`)
   }
-  return findAndReplaceVars(propToCheck)
-}
-
-function findAndReplaceVars(styleObject) {
-  return Object.keys(styleObject).reduce((prev, current) => {
-    const sanitizedVal = checkForNestedSelectors(styleObject[current])
-    return {
-      ...prev,
-      [current]: sanitizedVal,
-    }
-  }, styleObject)
 }
 
 const convertRules = (rules, res = {}) => {
@@ -52,53 +48,19 @@ const convertRules = (rules, res = {}) => {
   return result
 }
 
-export default function convertToJS(input) {
-  // Parse CSS string into rules array
-  try {
-    const parsedCss = css.parse(input)
-    const { rules } = parsedCss.stylesheet
-    const initial = convertRules(rules)
-
-    return applyCompositions(initial)
-  } catch (err) {
-    throw new Error(`Invalid CSS input: ${err}`)
-  }
+function findAndReplaceVars(styleObject) {
+  return Object.keys(styleObject).reduce((prev, current) => {
+    const sanitizedVal = checkForNestedSelectors(styleObject[current])
+    return {
+      ...prev,
+      [current]: sanitizedVal,
+    }
+  }, styleObject)
 }
 
-function applyCompositions(obj) {
-  for (const [className, value] of Object.entries(obj)) {
-    if (typeof value !== 'object') {
-      obj[className] = value
-      continue
-    }
-    const { composes, ...overrides } = value
-    obj[className] = {}
-    if (composes) {
-      if (composes in obj) {
-        deepMerge(obj[className], obj[composes])
-      } else if (composes.includes(' from ')) {
-        // Ignore here, will be appended in later step
-        deepMerge(obj[className], value)
-      } else {
-        composes.split(/\s+/).forEach((otherKey) => {
-          deepMerge(obj[className], obj[otherKey])
-        })
-      }
-    }
-    deepMerge(obj[className], overrides)
+function checkForNestedSelectors(propToCheck) {
+  if (typeof propToCheck === 'string') {
+    return propToCheck
   }
-  return obj
-}
-
-function deepMerge(source, target) {
-  for (const [key, value] of Object.entries(target)) {
-    // Overwrite primitive values, merge objects together
-    if (key in source && typeof value === 'object') {
-      deepMerge(source[key], value)
-    } else if (typeof value === 'object') {
-      source[key] = { ...value }
-    } else {
-      source[key] = value
-    }
-  }
+  return findAndReplaceVars(propToCheck)
 }
